refactor (step3): result 리패키징
