+++ src/main/kotlin/blackjack/domain/result/PlayersResult.kt
@@ -1,5 +1,8 @@
-package blackjack.domain
+package blackjack.domain.result
 
+import blackjack.domain.Dealer
+import blackjack.domain.Name
+import blackjack.domain.Players
 import blackjack.domain.state.GameResultState
 
 @JvmInline
